requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
future==0.18.2
graphviz==0.15
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pydot==1.4.1
pyimgur==0.6.0
pyparsing==2.4.7
requests==2.25.1
six==1.15.0
transitions==0.8.5
urllib3==1.26.2
Werkzeug==1.0.1
